Created a new view and added buttons/icons
Added a new view, acitivity_no_result_, and created a java class for it
as well. Added an activity node in android manifest with parent
MainActivity. Also, there is a new icon/button for drinks. The old one
is still there though in case we want to change. I removed the phone
button as well because we can’t get it to work properly.